feat(cli): introduce cobra root + app subcommand wrapping main bot
Combines plan Task 6 (root.go) + Task 7 (app.go extraction) since
they must land together to keep the build green.

- cmd/agentdock/root.go: rootCmd + version vars + Execute()
- cmd/agentdock/app.go: appCmd wrapping runApp(), which is the former
  main() body with os.Exit calls converted to returned errors. Also
  registers workerCmd with DisableFlagParsing so the existing
  runWorker() (which parses os.Args[2:] itself) continues to work.
- cmd/agentdock/adapters.go: extracted agentRunnerAdapter,
  repoCacheAdapter, slackPosterAdapter, parseLogLevel
- cmd/agentdock/main.go: gutted to 'func main() { Execute() }'
- go.mod: spf13/cobra promoted from indirect to direct
