+++ go.mod
@@ -7,6 +7,7 @@ require (
 	github.com/google/go-github/v60 v60.0.0
 	github.com/redis/go-redis/v9 v9.18.0
 	github.com/slack-go/slack v0.20.0
+	github.com/spf13/cobra v1.10.2
 	github.com/xuri/excelize/v2 v2.10.1
 	golang.org/x/sync v0.20.0
 	golang.org/x/term v0.42.0
@@ -21,7 +22,6 @@ require (
 	github.com/inconshreveable/mousetrap v1.1.0 // indirect
 	github.com/richardlehane/mscfb v1.0.6 // indirect
 	github.com/richardlehane/msoleps v1.0.6 // indirect
-	github.com/spf13/cobra v1.10.2 // indirect
 	github.com/spf13/pflag v1.0.9 // indirect
 	github.com/tiendc/go-deepcopy v1.7.2 // indirect
 	github.com/xuri/efp v0.0.1 // indirect
